SKip idle_timeout tests for ownership pool
